* feat: 增加系统管理员角色

管理员在后台可以对用户的账号进行操作。

* refactor: 重构文件存取服务接口与实现

* style: 删除冗余代码

- `ApiResponse` 未使用接口
- `NotificationEventListener` / `Profile` 冗余导入

* feat: 新增编辑部公告功能

系统管理员可进行公告管理，已发布公告可公开查看。

* feat(Announcement): 新增公告摘要和置顶功能

* fix: 图片文件MIME类型

* feat(FileStorageService): 改进文件存取

* feat: 增强异常处理信息

使用BussinessException提供更详细的错误信息

* fix: 不应对用户口令序列化

* feat: 添加公开接口以支持无需认证的公告查询

* feat(PublicController): 添加公开接口以查看公告图片

+++ src/main/java/top/tankimzeg/editorial_system/service/UserService.java
@@ -1,11 +1,14 @@
 package top.tankimzeg.editorial_system.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import top.tankimzeg.editorial_system.entity.User;
+import top.tankimzeg.editorial_system.exception.BusinessException;
 import top.tankimzeg.editorial_system.repository.UserRepo;
 
 import java.time.LocalDateTime;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -30,7 +33,10 @@ public class UserService {
         return userRepo.findById(id);
     }
 
-    public void deleteAuthor(Long id) {
+    public void deleteUserById(Long id) {
+        if (!userRepo.existsById(id)) {
+            throw new BusinessException(HttpStatus.NOT_FOUND, "用户不存在");
+        }
         userRepo.deleteById(id);
     }
 
@@ -42,4 +48,7 @@ public class UserService {
         return userRepo.existsUserByEmail(email);
     }
 
+    public List<User> getAllUsers() {
+        return userRepo.findAll();
+    }
 }
